Wednesday Checkin

+++ PizzaTracker/app/js/controllers/queueController.js
@@ -1,5 +1,6 @@
 ﻿app.controller('queueController', ['$scope', '$location', 'queueService', 'resourceService', function ($scope, $location, queueService, resourceService) {
     $scope.queue = {};
+    $scope.statuses = {};
 
     $scope.getQueue = function () {
         queueService.getQueue()
@@ -7,8 +8,7 @@
                 console.log("queue gotten success " + response);
                 $scope.queue = response;
                 for (var i = 0; i < $scope.queue.length; i++) {
-                    $scope.queue[i].Statuses = $scope.resources.Statuses;
-                    $scope.queue[i].Status = $.grep($scope.queue[i].Statuses, function (e) { return e.Id == $scope.queue[i].Status.Id; });
+                    $scope.queue[i].Status = lookupById($scope.statuses, $scope.queue[i].Status.Id);
                 }
             },
                 function (err) {
@@ -28,24 +28,28 @@
     };
 
     $scope.deleteQueue = function (id) {
-        queueService.deleteQueue(id)
-            .then(function (response) {
-                console.log("Queue delete success " + response);
-                $scope.getQueue();
-            },
-       function (err) {
-           console.log("Queue delete error " + err);
-       });
+        alertify.confirm("Are you sure you want to remove this Pizza from the Queue?", function (e) {
+            if (e) {
+                queueService.deleteQueue(id)
+                    .then(function (response) {
+                        console.log("Queue delete success " + response);
+                        $scope.getQueue();
+                    },
+               function (err) {
+                   console.log("Queue delete error " + err);
+               });
+            }
+        });
     };
 
     resourceService.getResources(false)
     .then(function (response) {
         console.log("resource gotten success " + response);
-        $scope.resources = response;
+        $scope.statuses = response.Statuses;
     },
         function (err) {
             console.log("resource gotten error " + err);
-            $scope.resources = {};
+            $scope.statuses = {};
         });
 
     $scope.getQueue();
